Reducer should not be an external
