--- src/state/statestore.spec.js
@@ -1,168 +1,0 @@
-/* eslint-env jest */
-
-import {reducer, store} from './statestore.js'
-
-describe('reducer', () => {
-  describe('initialisation', () => {
-    it('initialisaes with an empty token list', () => {
-      expect(reducer().tokens).toEqual([])
-    })
-
-    it('initialises with a null result', () => {
-      expect(reducer().result).toBe(null)
-    })
-  })
-
-  describe('number as action', () => {
-    const action = {
-      type: 'APPEND_NUMBER',
-      value: 0
-    }
-
-    describe('when the token list is empty', () => {
-      it('adds the number to the token list', () => {
-        const result = reducer(undefined, action).tokens[0]
-        expect(result.type).toBe('number')
-        expect(result.resolve().toString()).toBe('0')
-      })
-    })
-
-    describe('when the token list is terminated by a number', () => {
-      const start = reducer(undefined, {
-        type: 'APPEND_NUMBER',
-        value: 1
-      })
-
-      it('appends the number to the existing number', () => {
-        const result = reducer(start, action).tokens[0]
-        expect(result.type).toBe('number')
-        expect(result.resolve().toString()).toBe('10')
-      })
-    })
-
-    describe('when the token list is terminated by an operator', () => {
-      const first = reducer(undefined, {
-        type: 'APPEND_NUMBER',
-        value: 0
-      })
-      const second = reducer(first, {
-        type: 'APPEND_OPERATOR',
-        value: '+'
-      })
-
-      it('adds the number to the token list', () => {
-        const result = reducer(second, action).tokens
-        expect(result[0].type).toBe('number')
-        expect(result[0].resolve().toString()).toBe('0')
-        expect(result[1].type).toBe('operator')
-        expect(result[2].type).toBe('number')
-        expect(result[2].resolve().toString()).toBe('0')
-      })
-    })
-  })
-
-  describe('operator as action', () => {
-    const action = {
-      type: 'APPEND_OPERATOR',
-      value: '+'
-    }
-
-    describe('when the token list is empty', () => {
-      it('does not add the operator to the token list', () => {
-        expect(reducer(undefined, action).tokens).toEqual([])
-      })
-    })
-
-    describe('when the token list already has a number', () => {
-      const start = reducer(undefined, {
-        type: 'APPEND_NUMBER',
-        value: 0
-      })
-
-      it('adds the operator to the token list', () => {
-        const result = reducer(start, action).tokens[1]
-        expect(result.type).toBe('operator')
-      })
-    })
-
-    describe('when the token list is terminated by an operator', () => {
-      const first = reducer(undefined, {
-        type: 'APPEND_NUMBER',
-        value: 0
-      })
-      const second = reducer(first, {
-        type: 'APPEND_OPERATOR',
-        value: '+'
-      })
-
-      it('does not add the operator to the token list', () => {
-        const result = reducer(second, action).tokens
-        expect(result).toHaveLength(2)
-      })
-    })
-  })
-
-  describe('delete digit', () => {
-    const del = {type: 'CLEAR_DIGIT'}
-
-    it('does nothing when the store is empty', () => {
-      const state = reducer(undefined, del)
-      expect(state.tokens).toEqual([])
-    })
-
-    it('will remove a single number', () => {
-      const first = reducer(undefined, {type: 'APPEND_NUMBER', value: 1})
-      const state = reducer(first, del)
-      expect(state.tokens).toEqual([])
-    })
-
-    it('will remove the last digit of a number', () => {
-      const first = reducer(undefined, {type: 'APPEND_NUMBER', value: 2})
-      const second = reducer(first, {type: 'APPEND_NUMBER', value: 3})
-      const state = reducer(second, del)
-      expect(state.tokens[0].resolve().toString()).toBe('2')
-    })
-
-    it('will remove an operator', () => {
-      const first = reducer(undefined, {type: 'APPEND_NUMBER', value: 4})
-      const second = reducer(first, {type: 'APPEND_OPERATOR', value: '+'})
-      const state = reducer(second, del)
-      expect(state.tokens).toHaveLength(1)
-    })
-  })
-
-  describe('delete all', () => {
-    const del = {type: 'CLEAR_ALL'}
-
-    it('does nothing when the store is empty', () => {
-      const state = reducer(undefined, del)
-      expect(state.tokens).toEqual([])
-    })
-
-    it('will remove all digits', () => {
-      const first = reducer(undefined, {type: 'APPEND_NUMBER', value: 2})
-      const second = reducer(first, {type: 'APPEND_NUMBER', value: 3})
-      const state = reducer(second, del)
-      expect(state.tokens).toEqual([])
-    })
-  })
-})
-
-describe('store', () => {
-  it('returns the initial state', () => {
-    expect(store.getState()).toEqual({
-      tokens: [],
-      result: null
-    })
-  })
-
-  it('processes an action', () => {
-    store.dispatch({
-      type: 'APPEND_NUMBER',
-      value: 2
-    })
-    const state = store.getState()
-    expect(state.tokens).toHaveLength(1)
-    expect(state.result).toBe(2)
-  })
-})
