+++ src/view/index.spec.js
@@ -1,18 +1,22 @@
 /* eslint-env jest */
 import ReactDOM from 'react-dom'
-import {store} from '../state/statestore'
+import * as store from '../state/store'
 
 describe('index', () => {
-  it('iteracts with the store and React.render()', () => {
+  it('interacts with the store and React.render()', () => {
+    const myStore = store.createStore()
+    jest.spyOn(store, 'createStore').mockImplementation(() => myStore)
+    jest.spyOn(myStore, 'getState').mockImplementation(() => undefined)
+    jest.spyOn(myStore, 'dispatch').mockImplementation(() => undefined)
+    jest.spyOn(myStore, 'subscribe').mockImplementation(() => undefined)
+
     jest.spyOn(ReactDOM, 'render').mockImplementation(() => undefined)
-    jest.spyOn(store, 'getState').mockImplementation(() => undefined)
-    jest.spyOn(store, 'dispatch').mockImplementation(() => undefined)
-    jest.spyOn(store, 'subscribe').mockImplementation(() => undefined)
+
     require('./index')
 
-    expect(store.getState).toHaveBeenCalled()
-    expect(store.dispatch).not.toHaveBeenCalled()
-    expect(store.subscribe).toHaveBeenCalled()
+    expect(myStore.getState).toHaveBeenCalled()
+    expect(myStore.dispatch).not.toHaveBeenCalled()
+    expect(myStore.subscribe).toHaveBeenCalled()
     expect(ReactDOM.render).toHaveBeenCalled()
   })
 })
